Refactor save_to_disk
